update
- added example data

+++ app/src/main/java/de/taskmaster/model/binding/AddressEditorHandler.kt
@@ -21,9 +21,10 @@ class AddressEditorHandler(val view: PlaceEditor, context: Context) : ToggleEdit
 
         //TODO: check if fields are valid
 
+        //TODO: refactor I guess
         address.street = (view.getView(R.id.item_place_street) as TextView).text.toString()
-        address.number = (view.getView(R.id.item_place_number) as TextView).text.toString()
-        address.zipCode = (view.getView(R.id.item_place_zip) as TextView).text.toString()
+        address.number = (view.getView(R.id.item_place_number) as TextView).text.toString().toInt()
+        address.zipCode = (view.getView(R.id.item_place_zip) as TextView).text.toString().toInt()
         address.city = (view.getView(R.id.item_place_city) as TextView).text.toString()
 
         view.add(address)
